fixed unused imports, functions.
