+++ tests/integration/components/registries/update-dropdown/component-test.ts
@@ -5,46 +5,74 @@ import { setupMirage } from 'ember-cli-mirage/test-support';
 import { percySnapshot } from 'ember-percy';
 import { module, test } from 'qunit';
 
-import Registration from 'ember-osf-web/models/registration';
-import RegistrationModel from 'ember-osf-web/models/registration';
+import Registration, { RegistrationReviewStates } from 'ember-osf-web/models/registration';
+import { RevisionReviewStates } from 'ember-osf-web/models/revision';
 import { click, setupOSFApplicationTest } from 'ember-osf-web/tests/helpers';
-import RevisionModel from 'ember-osf-web/models/revision';
-import { assert } from '@ember/debug';
-import decaf from 'ember-osf-web/mirage/scenarios/registrations';
 
-// get all accepted revisions on a particular Registration; sort by date in descending order (most recent edits first)
-function getAllRevisionsTest(revision: ModelInstance<RevisionModel>){
-    return revision.models
-        .filter((revisions: { decaf: any; }) => revisions.decaf)
-        .sort((a,b) => parseInt(b.dateCreated) - parseInt(a.dateCreated)); // because only approved edits are created
+import { Permission } from 'ember-osf-web/models/osf-model';
+
+function allRevisions(registration: ModelInstance<Registration>){
+    return registration.revisions
+        .filter(revision => !revision.versionNumber)
+        .sort((a,b) => parseInt(b.id, 10) - parseInt(a.id, 10));
 }
 
-module('Acceptance | updateDropdown', hooks => {
-    setupEngineRenderingTest(hooks, 'registries');
-    setupIntl(hooks);
+module('Acceptance | update dropdown', hooks => {
     setupOSFApplicationTest(hooks);
     setupMirage(hooks);
 
-    test('update dropdown component', async assert => {
+    test('update dropdown - all revisions', async assert => {
+        const userServer = server.create('user', 'loggedIn');
+
+        const node = server.create('node', {
+            id: 'cobalt',
+            title: 'Outcome Reporting Testing',
+            currentUserPermissions: [Permission.Read],
+            currentUserIsContributor: true,
+            registration: true,
+            public: true,
+        });
+        server.create('registration', {
+            registeredFrom: node,
+            reviewsState: RegistrationReviewStates.Accepted,
+            revisionState: RevisionReviewStates.Approved,
+        });
+        server.create('draft-registration', {
+            branchedFrom: node,
+        }, 'currentUserIsReadAndWrite');
+        server.create('revision', {
+            reviewState: RevisionReviewStates.Approved,
+            versionNumber: 1,
+        });
+        server.create('revision', {
+            reviewState: RevisionReviewStates.Approved,
+            versionNumber: 2,
+        });
+        server.create('revision', {
+            reviewState: RevisionReviewStates.Approved,
+            versionNumber: 3,
+        });
+        const beforeCount = userServer.registrations.length;
 
         const revisions = server.createList('revision', 10, 'withChildren');
-        const beforeCount = revisions.length;
 
-        // TODO update to proper route name once on test
-        await visit('/decaf');
+        const revisions = allRevisions(registration);
 
-        // for the component
-        assert.dom('[data-test-verify-update-dropdown-element]').exists();
+        const url = `/${node.id}/registrations`;
+        await visit(url);
+
+        // for the component itself
+        assert.dom('[data-test-update-button]').exists();
         await percySnapshot(assert);
+
         // for the list of revisions
-        assert.dom('[data-test-verify-update-dropdown-list]').doesNotExist(); // TODO create data test link
+        assert.dom('[data-test-update-list]').doesNotExist();
         await percySnapshot(assert);
-
-        await click('[data-test-verify-update-dropdown-list]');
-        assert.dom('[data-test-verify-update-dropdown-list').exists();
-        decaf.reload();
-        assert.dom('[data-test-verify-update-dropdown-list]').doesNotExist();
-        assert.equal(revisions.length, beforeCount, 'Correct number of revisions');
-        assert.ok(revisions.every(revision => revision.dateCreated), 'All approved revisions present by date.');
+        await click('[data-test-update-list]');
+        assert.dom('[data-test-list-view]').exists();
+        assert.equal(registration.revisions.length, beforeCount, 'Correct number of revisions');
+        assert.ok(revisions.every(revision => revision.versionNumber), 'Version number present.');
+        userServer.reload();
+        assert.dom('[data-test-list-view]').doesNotExist();
     });
- });
+});
